update unit test

+++ gridgo-framework/src/test/java/io/gridgo/framework/test/MessageUnitTest.java
@@ -20,25 +20,25 @@ public class MessageUnitTest {
 	public void testMessage() {
 		var msg1 = Message.of(Payload.of(BObject.ofEmpty().setAny("key", 1), BValue.of(1)));
 		msg1.getPayload().addHeaderIfAbsent("key", 2);
-		Assert.assertEquals(1, msg1.getPayload().getHeaders().getInteger("key"));
+		Assert.assertEquals(1, msg1.getPayload().getHeaders().getInteger("key").intValue());
 		msg1.getPayload().addHeaderIfAbsent("test", 2);
-		Assert.assertEquals(2, msg1.getPayload().getHeaders().getInteger("test"));
+		Assert.assertEquals(2, msg1.getPayload().getHeaders().getInteger("test").intValue());
 		msg1.getPayload().addHeader("test", 1);
-		Assert.assertEquals(1, msg1.getPayload().getHeaders().getInteger("test"));
+		Assert.assertEquals(1, msg1.getPayload().getHeaders().getInteger("test").intValue());
 		msg1.attachSource("test");
 		Assert.assertEquals("test", msg1.getMisc().get(MessageConstants.SOURCE));
 		msg1.setRoutingIdFromAny(1);
-		Assert.assertEquals(1, msg1.getRoutingId().get().getInteger());
+		Assert.assertEquals(Integer.valueOf(1), msg1.getRoutingId().get().getInteger());
 		var msg2 = Message.of(BValue.of(1), Payload.of(null));
-		Assert.assertEquals(1, msg2.getRoutingId().get().getInteger());
+		Assert.assertEquals(Integer.valueOf(1), msg2.getRoutingId().get().getInteger());
 		var msg3 = Message.of(BValue.of(1), Collections.singletonMap("test", 1),
 				Payload.of(null));
 		Assert.assertEquals(1, msg3.getMisc().get("test"));
 		var msg4 = Message
 				.parse(BArray.ofEmpty().addAny(1).addAny(BObject.ofEmpty().setAny("test", 1)).addAny(1));
-		Assert.assertEquals(1, msg4.getPayload().getId().get().getInteger());
-		Assert.assertEquals(1, msg4.getPayload().getHeaders().getInteger("test"));
-		Assert.assertEquals(1, msg4.getPayload().getBody().asValue().getInteger());
+		Assert.assertEquals(Integer.valueOf(1), msg4.getPayload().getId().get().getInteger());
+		Assert.assertEquals(Integer.valueOf(1), msg4.getPayload().getHeaders().getInteger("test"));
+		Assert.assertEquals(Integer.valueOf(1), msg4.getPayload().getBody().asValue().getInteger());
 	}
 
 	@Test
@@ -53,11 +53,11 @@ public class MessageUnitTest {
 
 	private void assertMessage(MultipartMessage msg) {
 		Assert.assertEquals(true, msg.getPayload().getHeaders().getBoolean(MessageConstants.IS_MULTIPART));
-		Assert.assertEquals(2, msg.getPayload().getHeaders().getInteger(MessageConstants.SIZE));
+		Assert.assertEquals(2, msg.getPayload().getHeaders().getInteger(MessageConstants.SIZE).intValue());
 		Assert.assertTrue(msg.getPayload().getBody().isArray());
 		Assert.assertEquals(2, msg.getPayload().getBody().asArray().size());
-		Assert.assertEquals(1, msg.getPayload().getBody().asArray().getArray(0).getObject(1).getInteger("key"));
-		Assert.assertEquals(2, msg.getPayload().getBody().asArray().getArray(1).getObject(1).getInteger("key"));
+		Assert.assertEquals(1, msg.getPayload().getBody().asArray().getArray(0).getObject(1).getInteger("key").intValue());
+		Assert.assertEquals(2, msg.getPayload().getBody().asArray().getArray(1).getObject(1).getInteger("key").intValue());
 		Assert.assertEquals(1, msg.getPayload().getBody().asArray().getArray(0).getInteger(2));
 		Assert.assertEquals(2, msg.getPayload().getBody().asArray().getArray(1).getInteger(2));
 	}
@@ -65,11 +65,11 @@ public class MessageUnitTest {
 	@Test
 	public void testPayload() {
 		var payload = Payload.of(BValue.of(1), BValue.of(2));
-		Assert.assertEquals(1, payload.getId().get().getInteger());
-		Assert.assertEquals(2, payload.getBody().asValue().getInteger());
+		Assert.assertEquals(Integer.valueOf(1), payload.getId().get().getInteger());
+		Assert.assertEquals(Integer.valueOf(2), payload.getBody().asValue().getInteger());
 		payload.setIdFromAny("test");
 		Assert.assertEquals("test", payload.getId().get().getString());
 		payload.setBody(BObject.ofEmpty().setAny("test", 1));
-		Assert.assertEquals(1, payload.getBody().asObject().getInteger("test"));
+		Assert.assertEquals(1, payload.getBody().asObject().getInteger("test").intValue());
 	}
 }
